Bring back to life the Spring tests that were left out for Java 9+

+++ jqm-all/jqm-tests/jqm-test-spring-1/src/main/java/com/enioka/jqm/test/spring1/domain/OfferingRepository.java
@@ -1,10 +1,8 @@
 package com.enioka.jqm.test.spring1.domain;
 
-import java.math.BigInteger;
-
 import org.springframework.data.repository.CrudRepository;
 
-public interface OfferingRepository extends CrudRepository<Offering, BigInteger>
+public interface OfferingRepository extends CrudRepository<Offering, Long>
 {
     Offering findByOfferingNumber(Long offeringNumber);
 }
